* master: (145 commits)
  [WIP]fix issues #4972 #4973 (#4972)
  bugfix `refresh maxAllowRunningJobs`
  bugfix NPE
  bugfix `head of empty list` again
  Bump axios from 0.21.4 to 1.6.0 in /linkis-web
  Bump axios from 1.5.0 to 1.6.0 in /linkis-web-next
  Fix integration test bug
  rollback bugfix `head of empty list`
  bugfix `head of empty list`
  Fix integration test bug
  add config
  feat: optimize process of closing spark job on k8s (#4957)
  add import `org.apache.spark.sql.api.python.*` (#4961)
  Fix integration test bug
  update action (#4962)
  Remove `linkis.entrance.auto.clean.dirty.data.enable` configuration
  Remove redundant configuration
  Keyword uppercase
  Bump postcss from 8.4.29 to 8.4.31 in /linkis-web-next
  refactor of linkis-web
  ...

# Conflicts:
#	.github/actions/chart-testing-action

+++ linkis-engineconn-plugins/spark/src/main/java/org/apache/linkis/engineplugin/spark/client/deployment/KubernetesApplicationClusterDescriptorAdapter.java
@@ -40,7 +40,7 @@ import org.slf4j.LoggerFactory;
 
 public class KubernetesApplicationClusterDescriptorAdapter extends ClusterDescriptorAdapter {
   private static final Logger logger =
-      LoggerFactory.getLogger(KubernetesOperatorClusterDescriptorAdapter.class);
+      LoggerFactory.getLogger(KubernetesApplicationClusterDescriptorAdapter.class);
 
   protected SparkConfig sparkConfig;
   protected KubernetesClient client;
@@ -66,12 +66,13 @@ public class KubernetesApplicationClusterDescriptorAdapter extends ClusterDescri
         .setJavaHome(sparkConfig.getJavaHome())
         .setSparkHome(sparkConfig.getSparkHome())
         .setMaster(sparkConfig.getK8sMasterUrl())
-        .setDeployMode(sparkConfig.getDeployMode())
+        .setDeployMode("cluster")
         .setAppName(sparkConfig.getAppName())
         .setVerbose(true);
     this.driverPodName = generateDriverPodName(sparkConfig.getAppName());
     this.namespace = sparkConfig.getK8sNamespace();
     setConf(sparkLauncher, "spark.app.name", sparkConfig.getAppName());
+    setConf(sparkLauncher, "spark.ui.port", sparkConfig.getK8sSparkUIPort());
     setConf(sparkLauncher, "spark.kubernetes.namespace", this.namespace);
     setConf(sparkLauncher, "spark.kubernetes.container.image", sparkConfig.getK8sImage());
     setConf(sparkLauncher, "spark.kubernetes.driver.pod.name", this.driverPodName);
@@ -111,6 +112,7 @@ public class KubernetesApplicationClusterDescriptorAdapter extends ClusterDescri
     addSparkArg(sparkLauncher, "--num-executors", sparkConfig.getNumExecutors().toString());
     addSparkArg(sparkLauncher, "--principal", sparkConfig.getPrincipal());
     addSparkArg(sparkLauncher, "--keytab", sparkConfig.getKeytab());
+    addSparkArg(sparkLauncher, "--py-files", sparkConfig.getPyFiles());
     sparkLauncher.setAppResource(sparkConfig.getAppResource());
     sparkLauncher.setMainClass(mainClass);
     Arrays.stream(args.split("\\s+"))
@@ -164,12 +166,12 @@ public class KubernetesApplicationClusterDescriptorAdapter extends ClusterDescri
     return client.pods().inNamespace(namespace).withName(driverPodName).get();
   }
 
-  public String getSparkDriverPodIP() {
+  public String getSparkUIUrl() {
     Pod sparkDriverPod = getSparkDriverPod();
     if (null != sparkDriverPod) {
       String sparkDriverPodIP = sparkDriverPod.getStatus().getPodIP();
       if (StringUtils.isNotBlank(sparkDriverPodIP)) {
-        return sparkDriverPodIP;
+        return sparkDriverPodIP + ":" + this.sparkConfig.getK8sSparkUIPort();
       } else {
         logger.info("spark driver pod IP is null, the application may be pending");
       }
@@ -194,12 +196,16 @@ public class KubernetesApplicationClusterDescriptorAdapter extends ClusterDescri
   @Override
   public void close() {
     logger.info("Start to close job {}.", getApplicationId());
+    client.close();
+    if (isDisposed()) {
+      logger.info("Job has finished, close action return.");
+      return;
+    }
     PodResource<Pod> sparkDriverPodResource =
         client.pods().inNamespace(namespace).withName(driverPodName);
     if (null != sparkDriverPodResource.get()) {
       sparkDriverPodResource.delete();
     }
-    client.close();
   }
 
   @Override
